Use EEPROM esp32 NVS instead of SPIFFS
